cleanup

+++ prikol/src/main/java/com/ydo4ki/fw/internal/lib/jlib/data/JByteFw.java
@@ -84,7 +84,7 @@ public final class JByteFw {
     }
 
     public static Byte unwrap(Val val) {
-        return (Byte) val._UNPACK_();
+        return val._UNPACK_();
     }
 }
