+++ prikol/src/main/java/com/ydo4ki/fw/internal/lib/jlib/data/JFloatFw.java
@@ -45,7 +45,7 @@ public final class JFloatFw {
     }
 
     public static Float unwrap(Val val) {
-        return Float.intBitsToFloat((Integer) val._UNPACK_());
+        return Float.intBitsToFloat(val._UNPACK_());
     }
 }
